[add]module PointAvailable
